checking changes for shared usecase. Fixed RPC calls.

+++ src/athena/server/posix_io.h
@@ -10,8 +10,8 @@ namespace athena {
 int posix_open(std::string filename, int mode, int flags);
 int posix_close(int fd);
 off_t posix_lseek(int fd, off_t offset, int whence);
-ssize_t posix_write(int fd, std::string buf, int count);
-std::string posix_read(int fd, int count);
+ssize_t posix_write(int fd, std::string buf, size_t count);
+std::string posix_read(int fd, size_t count);
 }  // namespace athena
 
 #endif  // ATHENA_POSIX_IO_H
